# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for PetSys application.",
        "title": "PetSys",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Daniel G_A",
            "url": "https://tec.mx/es",
            "email": "A01794498@tec.mx"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/v1/pets": {
            "get": {
                "description": "Get pets info by user",
                "tags": [
                    "pet"
                ],
                "summary": "Get pets info by user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_pet.Pets"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Save a new user pet",
                "tags": [
                    "pet"
                ],
                "summary": "Save Pet",
                "parameters": [
                    {
                        "description": "Save request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_pet.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_pet.GeneralResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pets/{petID}": {
            "get": {
                "description": "Gets pet information",
                "tags": [
                    "pet"
                ],
                "summary": "Gets pet information by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pet ID",
                        "name": "petID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_pet.GeneralResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Updates multiple fields from pet in the DB",
                "tags": [
                    "pet"
                ],
                "summary": "Updates pet information",
                "parameters": [
                    {
                        "description": "Save request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_pet.UpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "pet ID",
                        "name": "petID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_pet.GeneralResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pets/{userID}": {
            "delete": {
                "description": "Deletes pet by ID",
                "tags": [
                    "pet"
                ],
                "summary": "Deletes a pet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/users": {
            "get": {
                "description": "Gets users saved",
                "tags": [
                    "user"
                ],
                "summary": "Gets all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.Users"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/users/login": {
            "post": {
                "description": "Logins a user and returns a JWT",
                "tags": [
                    "user"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.LoginResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/users/signup": {
            "post": {
                "description": "Register a User with email and password",
                "tags": [
                    "user"
                ],
                "summary": "SignUp User",
                "parameters": [
                    {
                        "description": "Register request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.RegisterResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/users/{userID}": {
            "get": {
                "description": "Gets user info with ID",
                "tags": [
                    "user"
                ],
                "summary": "Gets user info by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Updates multiple fields from user in the DB",
                "tags": [
                    "user"
                ],
                "summary": "Updates user information",
                "parameters": [
                    {
                        "description": "Save request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.UpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MNA-project_pkg_internal_user.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Deletes user by ID",
                "tags": [
                    "user"
                ],
                "summary": "Deletes user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "MNA-project_pkg_internal_pet.GeneralResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "owner_name": {
                    "type": "string"
                },
                "register_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_pet.Pets": {
            "type": "object",
            "properties": {
                "pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MNA-project_pkg_internal_pet.GeneralResponse"
                    }
                }
            }
        },
        "MNA-project_pkg_internal_pet.RegisterRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_pet.UpdateRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "last_name",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.RegisterResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.UpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "MNA-project_pkg_internal_user.Users": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MNA-project_pkg_internal_user.Response"
                    }
                }
            }
        }
    }
}